Fixing remaining javadocs, package change on tests, fixing converter issues
